+++ lib/rails_health_monitor.rb
@@ -1,18 +1,20 @@
 # frozen_string_literal: true
 
 require_relative "rails_health_monitor/version"
+require_relative "rails_health_monitor/checks/database"
 require "socket"
 
 module RailsHealthMonitor
   class Error < StandardError; end
-  
+
   def self.check
     {
       ruby_version: RUBY_VERSION,
       rails_version: defined?(Rails) ? Rails.version : nil,
       environment: defined?(Rails) ? Rails.env : nil,
       hostname: Socket.gethostname,
-      timestamp: Time.now.utc
+      timestamp: Time.now.utc,
+      database: RailsHealthMonitor::Checks::Database.call
     }
   end
 end