+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## [0.1.0] - 2026-06-22
 
 - Initial release
+
+## 0.1.1
+
+### Added
+- Database health check.
+- Improved health check response structure.